com.unity.memoryprofiler@0.1.0-preview.7 ## [0.1.0-preview.7] - 2019-08-02 ### Added  - Added Screenshots getting taken on capture for Unity versions starting from 2019.3 and up.  - Added window tab icon.  - Added a popup warning about the potential for sharing personally identifying or otherwise sensitive data when sharing snapshots.  - Added info to the documentation regarding potential sensitive data contained in snapshot files.  - Added a button to the preferences to reset the opt-out decisions for above mentioned warning popup.  - Added byte size formatting for size columns in the tables.  - Added "Open Folder" option to the snapshot options menu.
### Changed
 - Fixed snapshot file rename functionality ([case 1131905](https://issuetracker.unity3d.com/product/unity/issues/guid/1131905/)).
 - Fixed snapshot file and meta data fields overlapping in 2019.3.x.
 - Fixed new snapshot folders getting created with every character change to the path in the Memory Profiler preferences ([case 1162851](https://issuetracker.unity3d.com/product/unity/issues/guid/1162851/)).
 - Fixed a numerical overflow when parsing snapshots with large amounts of objects and/or allocations.
 - Fixed compile issues due to the removed "EditorApplication.scriptingRuntimeVersion" API.
 - Fixed the target selection drop-down which stopped working in 2019.3
 - Moved the Memory Profiler preferences under Analysis/Memory Profiler in the Preferences window to group it with Profiler Window Settings.
 - Fixed alternating table row colors as well as row selection in 2019.3
 - Fixed link text color being hard to read in the light editor skin.

### Removed
 - Metadata collection for Scripting Runtime version for Unity >= 2019.3 since "EditorApplication.scriptingRuntimeVersion" was removed.

+++ Editor/Profiler/ManagedDataCrawler.cs
@@ -353,7 +353,7 @@ namespace Unity.MemoryProfiler.Editor
                 CrawlDataStack = new Stack<StackCrawlData>();
 
                 TypesWithStaticFields = new List<int>();
-                for (int i = 0; i != snapshot.typeDescriptions.Count; ++i)
+                for (long i = 0; i != snapshot.typeDescriptions.Count; ++i)
                 {
                     if (snapshot.typeDescriptions.staticFieldBytes[i] != null
                         && snapshot.typeDescriptions.staticFieldBytes[i].Length > 0)
@@ -474,7 +474,7 @@ namespace Unity.MemoryProfiler.Editor
 
         static void AddupRawRefCount(CachedSnapshot snapshot)
         {
-            for (int i = 0; i != snapshot.connections.Count; ++i)
+            for (long i = 0; i != snapshot.connections.Count; ++i)
             {
                 int iManagedTo = snapshot.UnifiedObjectIndexToManagedObjectIndex(snapshot.connections.to[i]);
                 if (iManagedTo >= 0)
